Separated shared templates from version-specific templates.
Templates for the TV version are mostly TV specific, but the other
versions share most of their templates, so I re-arranged the
definitions of the templates in the index files in two sections:
- a templates section that is common to phone, tablet and desktop,
and can safely undergo a copy-and-paste.
- a templates section that is specific to each version and should
not be copied over to another version.

Note the couple of specificities that remain can probably be
reduced to nill. That would be a good idea to simplify maintenance
in the future.