Support for mixed content feeds added with a DynamicContainer view.
The DynamicContainer view selects the right view to use to render
the underlying model based on the model to render. In other words,
the detail page uses the right view for the type of content to
render (image, video, news article, twitter message, etc.)

A couple of details pages still need to be specified and feeds
with mixed content should use the same layout for all items in the
feed.