Dropped iScroll in sound view as it may conflict with audio player.
On Android Ice Cream Sandwich, the audio player of the browser
fails to respond to the click event on the play button when iScroll
is activated.

There must be some kind of conflict going on, perhaps due to CSS
transforms and/or overlays. The detail view for music should not
require scrolling, so I deactivated iScroll in these views.