Manifest fiel: fixed array definition. Beware of the items property!
The "items" property defines the syntax that array items need to
follow. But beware, if it's an array, it sets the syntax for each
array item at the position of the syntax in the array. In short,
do not use an array for the "items" property. (see RFC for more
details).