List views not re-built anymore. Scroller refreshed when view is shown.
The listing view is no longer re-created unless the list of items
to render has changed. The created DOM fragment is re-used directly
and displayed once more.

The update on the Joshfire Framework fixes a bug with iScroll that
may incorrectly refresh the size of scrollers that are not displayed
when "resize" events are triggered. Since the dimensions of the
viewport may have changed (e.g. following a change in the orientation
of the screen) while a view was hidden, it is a good idea to force
a refresh of the scroller when a view is displayed in any case.

Tested on iPhone and Android. Still pretty slow on Android phones.