Data no longer reloaded if already there when tab is displayed.
The phone version was the only one not to reload data when a tab
was displayed. This has now been extended to all versions to speed
up rendering on all screens.

Note there is no "refresh" button on tablets/tv/desktop, so no way
to refresh the data displayed other than by restarting the app
(even then, if the datasource was shipped with a local cache, the
local version gets used if it hasn't expired).