Right thumbnail URL used in all listing views.
The feature was already there but there was a bug so the large
image was always displayed.

I also added a widthRatio parameter to the getThumbnail function
that takes a number between 0 and 1 to select the right thumbnail.
Typically, 0.2 means 20% of the screen width and is good for most
listing.

For pictures, I sticked to 0.2 although 0.5 would be closer to the
truth. The image quality looks acceptable with 0.2.

Note the use of a lower quality for images speeds up loading a bit
but that does not speed up scrolling as far as I can tell.