Initial view shown when ready. Screen width saved to avoid flickering.
The initial view was rendered as soon as it was created but was not
yet properly initialized. Content is now loaded before that view is
rendered.

The getThumbnail function that retrieves the right thumbnail based
on the dimensions of the screen caused the screen to flicker on iOS
devices because it accessed document.body.clientWidth that seems to
force a refresh of the screen. That width is now stored once and for
all when the application is initialized.