Fade-in effect added to detail view for news article images.
The detailed view now uses the newly created ImagesLoader
class to extract images and display a spinner while it loads them
in the background.

Note that, to avoid flickering on iOS, that means we're forcing
more and more elements down to the GPU, which means we might end
up with memory problems.