Generic spinner displayed while all images load in the background.
The code uses the "loading" class set by the view while its content
is being loaded. The loader automatically disappears when content
is loaded thanks to the "load" event triggered (and handled) by the
View class.

Data loading now uses a dedicated "dataloading" classname not to
conflict with that of the view assets loading.