Video detail view fixes. Various CSS improvements for articles.
All versions (save TV that behaves differently) now let the user
scroll the bottom portion of the screen in the detailed view of
a video content. The room available for scrolling is kind of small,
but there is no real way to prevent the likely iframe that contains
the video to handle touch events, which effectively makes the whole
app scroll and not only the right part in tablet version.

The commit also introduces a number of small CSS improvements to
display articles in a news feed, e.g. to ensure that images don't
exceed the width available and to add a left margin to lists.