Improved CSS rules for HTML content
The HTML content included in the "articleBody" or "description"
property of an item is now wrapped in an "article" div container.

Main CSS changes are for headers. Sleek does not exactly use headers
as it should, meaning that the title of a detailed view is typically
an <h4> tag whereas it should rather be an <h2> tag. We're also
using <h2> and <h3> for the date and/or author name which are not
headers per se. Anyway, HTML within articles should now render
correctly even it includes h1/h2 tags.