Made Photoswipe react to "click" instead of "touchstart" for toolbar
Touch events are poorly implemented in the default Android browser.
Photoswipe reacts on "touchstart" by default to handle clicks on
the close button in the toolbar and closes the view accordingly.

The problem is that there is no way to prevent the following "click"
event on Android that gets triggered on the element of the app view
behind the close button, which happens to be the button that leads
to the first tab in the phone version.

I slightly updated Photoswipe not to react on "touchstart". I added
a top comment that explains the difference between that version and
a pure vanilla Photoswipe library to ease updates.