Forced body margin to 0 to fix initial spinner on Webkit browsers
The initial spinner is embedded in the HTML with a couple of styles
so that it appears right away while the rest of the assets loads in
the background. The styles did not reset body margins though and
for some reason I don't quite get, that margin affects the rendering
of the animated gif rendered as a background image at the center
of the body.

Setting the body margin to 0 solved the problem.