Home page: explicit redirection to URL of first tab
The "home" route controller redirects the user to the first tab in
the application. It used to call the controller of the first tab
directly without updating the URL in the location bar.

Unfortunately, the "activate" function of the Toolbar view, used in
the TV version, goes beyond its rights in updating the URL itself
to match the URL of the selected item in the toolbar. That update
caused the controllers to run twice to display the home page in
the TV version.

In turn, that meant the datasource of the first tab was fetched twice
in a row, possibly in parallel. That's a problem because one cannot
call "find" on a datasource in parallel for the time being (missing
"feature" on data.js for the time being).

Explicitly redirecting the user to the URL of the first tab fixes
the issue.