HTML templates for products and photos expected ImageLoader
The code used ImagesLoader (plural form) but HTML templates for
products and photos (when in a mixed feed such as a Tumblr feed)
actually expected the presence of "imageUrl", set by ImageLoader.
I updated the code accordingly.

Also moved the code that triggered "load" hooks to before call to
initial "render", in case "render" works synchronously (not the
case given the views involved but that might happen).