Fixed "Refresh" link in phone version.
The "refresh" link always went back to the first tab of the application
because of the "link accelerator" added recently that listens to
"touchstart" and "mousedown" events instead of clicks. The refresh link
was defined in the markup with href="#" which made the application
move back home. I removed the href attribute.