"Load more" improved. Framework updated. Woodman logs added.
Switched to the latest version of the framework that handles the
"load more" feature more generically with a "loading more" class
name. and trigger the appropriate events on the collection itself
instead of on the List view.

The "app/woodmanConfig.js" file contains a default Woodman
configuration that activates logs on the framework as well. This is
for debugging purpose. All references to Woodman should disappear
when the template is built.

CSS and template not yet built.

Two main bugs remain:
- There are a couple of hiccups on the grid and vertical list views
in the TV version when new items get added to the lists
- The use of DynamicContainer means that list views are created
twice internally. It does not impact the app from a visual perspective
although this may be the cause for the flickering on iPad devices. It
does impact the performance of the app somehow (the ghost views
respond to the events triggered by the underlying collection).