Switched to latest version of the framework (rendered, update, remove)
The latest version of the framework brings a better tracking of
views, in particular for the DynamicContainer view.

Things that still do not work well:
- The Samsung version is borked. More fixes needed. The code has not
been updated to drop the "<li>" level to start with and the specific
"collection" needs to derive from "joshlib!collection".
- The flickering effect should still be around. Ghost views are not
kept around anymore, which is good, but they are still created
temporarily. It should be easier to track things down now that we
have better logs.