Samsung TV image collection now derives from joshlib!collection
The specific image collection is useful to filter out images that
are too large as Samsung TV models tend to crash when requested to
render such images.

The collection used to derive from Backbone.Collection directly
but the "joshlib!collection" class now adds useful logic to fetch
more items in a collection and to trigger load events when new items
are fetched.