Improved empty feed support. Ensured views are rendered only once
The commit brings in the latest version of the framework that
provides a more consistent update mechanism across views that do not
force a full refresh when not needed.

Sleek now renders views once when they get created. Updates made to
the models and collections bound to the views trigger the additional
magic needed to update the content displayed. This is probably not
enough to fix the flickering effect on iPads, but that should still
improve things and avoid useless updates that could happen in the
past.

Samsung TV version still not operational.