Fixed Samsung version (to be tested on actual devices!)
The Samsung TV version now derives from the TV version of Sleek to
maximize the amount of code that is shared across versions, as most
of the code of the Samsung TV version was actually a copy and paste
of the code of the default TV version.

The "load more" feature should now also work on the Samsung version.

I haven't checked the code on an actual device. While the update does
not bring anything "worrysome" in the sense that the updates do not
change a lot from a rendering perspective, history shows that the
template is likely still broken in practice ;)