Fixed startup crash on desktop in optimized version of the template
For some reason, there is a difference in behavior between the non
compiled version and the compiled version of the "onReady" utility
toolbelt. In the compiled version on desktops, it sends in the
underyling event as first parameter.

The Sleek.initialize function takes a potential callback function as
first parameter. The template crashed when it received the event
object. I updated the code to ensure that the initialize function
never receives the event object.

Related bug in Pivotal:
https://www.pivotaltracker.com/story/show/47715579