Ensured sections are created in a "hidden" state.
During application startup, all the sections were displayed at once
one on top of the others with a spinner up until the first route was
executed.

Also switched to the latest version of the framework that handles the
"hidden" initialization option and that also does not update the DOM
for nothing when the code calls "show/hide" whereas hidden state is
already correct.