Single video datasource: fixed available width/height computation
The "getContentWidth" and "getContentHeight" functions assumed that
all tabs were composed of a list and detail views but that is no
longer the case as some of the tabs may only contain a detail view.

The code of these functions had to check the dimensions of the
detail view because, many commits before the latest one, the
width of a slide panel in the framework was not the width of the
viewport but the sum of all the widths of the panels it contained.
That is no longer the case, we can now simply check the dimensions
of the container element to determine the available width and height.