Fixed flickering effect on iPad retina when loading/switching tabs
The flickering effect occurred because the toolbar gets rendered more
than once for some reason. In the end, the solution was to force the
rendering of the toolbar through the GPU.

I took the opportunity of the change to simplify the rendering of the
toolbar a bit as it was using a combinaison of floats and absolute
positioning that was not really necessary.

The utils.less now features a ".accelerate" shortcut. Note that
setting "backface-visibility" to "hidden" is not enough anymore
to force an element through the GPU on iPad retina devices. The
shortcut gets replaced by a combinaison of "backface-visibility",
"perspective" and "translateZ". They should not all be needed but
that should not have any drawback in practice.