Do not show toolbar when there is only one datasource (bis).
Since the new update to the view and the introduction of the hidden
property, this was broken. Indeed, the view now applies directly a
jQuery.show to its element. We used a "display:none" in the css
to hide the toolbar when needed.

To circumvent that, we now use "visibility:hidden".

This should fix https://www.pivotaltracker.com/story/show/29702415 for good.