"Share" icon displayed in detail page in phone version
The "share" icon gets displayed when the user accesses a details page,
it allows to share the item.

The "share" icon is displayed in the topbar at the same position as
the "refresh" icon.

In a proper implementation, the "share" icon would only be displayed
if there is a "share" add-on. Clicking on that icon would start the
"share" activity. Right now, add-ons are not checked and the hook is
not triggered. The action needs to be hardcoded in the "share"
function of "sleek.phone.js".

Notes:
- the icon is a copy of the "refresh" icon for the time being,
pending a better icon.
- the case when a datasource contains only one item is not supported,
meaning the "refresh" icon will be displayed in that case, not the
"share" icon.
- the case when the item cannot be shared (static local content) is
not supported either, meaning the "share" icon is always displayed.