The type of a list item now has priority over the type of the list
The list may contain mixed content. When that happens, the type of the
list is flagged as "other" and list items only display the title of
the underlying item. Using the type of the item makes the code use
the image, description and appropriate template based on the type
of each item.

Related bug in Pivotal:
https://www.pivotaltracker.com/story/show/55064818