Lists displayed items more than once
Due to the latest update of the framework and to an unfortunate
collision between the DynamicContainer class and the List class,
items in a list were rendered twice (save the first one).

The bug occurred because the List class now listens to "add" events
on the collection (for "fetchMore" purpose) and renders new items
accordingly. The DynamicContainer also listened to "add" events and
typically created the List when it received the first "add" event.
That List was initialized with all the items in the collection but
since it received further "add" events, it thought that new items
were being added to the collection.

Related bug in Pivotal:
https://www.pivotaltracker.com/story/show/55094146