Grid view: previously selected item not "unselected" on Samsung 2011
The bug occurred because the "dummy putInnerHTML" workaround had been
removed from the optimized HTML of the Samsung version of Sleek. The
workaround is there to force a call to "widgetAPI.putInnerHTML"
whenever a user action is detected (in other words whenever the user
presses a key on the remote). While the call should normally be
totally useless, it forces a DOM refresh that the browser may
otherwise miss when jQuery methods such as "addClass" or "removeClass"
are called.

Related bug in Pivotal:
https://www.pivotaltracker.com/story/show/55063830