Added support Dailymotion Flash player on Samsung Smart TVs
Youtube and Dailymotion players are now supported by a common
"VideoOverlay" view. The code looks at the URL of the video to play
to determine the player to use.

Successfully tested on a Samsung 2012 TV model, although note that
the TV does not always appreciate adjustments made to the playback
quality by the player: the audio disappears or becomes noise when
that happens. There is unfortunately no way to force the playback
quality to a particular setting programmatically. One may only set
a "suggested" playback quality that the player typically uses as
initial level and modifies after a few seconds:
http://www.dailymotion.com/fr/doc/api/player/javascript_api

The player does not run on 2011 models, probably because Flash 9 is
required and not available on these models.

Related story in Pivotal:
https://www.pivotaltracker.com/story/show/55474238